fix build when satis.json is not in default location
In SatisController, the Process used to invoke the `satis build`
command has its arguments passed in the $input parameter.
When the processed is invoked, these are not passed properly
and they are effectively ignored: the command that is launched is
`bin/satis` with no arguments.

Adding the arguments to the first parameter fixes this.