Update README to include more detail on using webhooks
We only really discovered how webhooks work by looking deep into the code. Before this we were going to use bitbucket pipelines to try to call satisfy, and were thinking because of the way the admin login works we'd need to hack it to create a custom url to kick off a build with some kind of pre-shared key auth. The whole time this wasn't needed as webhooks work perfectly.